+++ ta4j/src/main/java/org/ta4j/core/indicators/TripleEMAIndicator.java
@@ -48,4 +48,9 @@ public class TripleEMAIndicator extends CachedIndicator<Decimal> {
         EMAIndicator emaEmaEma = new EMAIndicator(emaEma, timeFrame);
         return Decimal.THREE.multipliedBy(ema.getValue(index).minus(emaEma.getValue(index))).plus(emaEmaEma.getValue(index));
     }
+    
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + " timeFrame: " + timeFrame;
+    }
 }