+++ ta4j/src/main/java/org/ta4j/core/indicators/volume/VWAPIndicator.java
@@ -70,4 +70,9 @@ public class VWAPIndicator extends CachedIndicator<Decimal> {
         }
         return cumulativeTPV.dividedBy(cumulativeVolume);
     }
+    
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + " timeFrame: " + timeFrame;
+    }
 }